Reddit_Crawler_code/requirements.txt
pymongo==3.11.4
requests==2.26.0
faktory==1.0.0
python-dotenv==0.19.0
